+++ nikiforov.alexander/T2/dataStruct.cpp
@@ -2,6 +2,7 @@
 #include <iomanip>
 #include "typesKeys.hpp"
 #include "iofmtguard.hpp"
+#include "delimiter.hpp"
 
 namespace nikiforov
 {
