+++ nikiforov.alexander/T2/delimiter.cpp
@@ -1,0 +1,20 @@
+#include "delimiter.hpp"
+
+namespace nikiforov
+{
+  std::istream& operator>>(std::istream& in, DelimiterIO&& dest)
+  {
+    std::istream::sentry sentry(in);
+    if (!sentry)
+    {
+      return in;
+    }
+    char c = '0';
+    in >> c;
+    if (in && (c != dest.exp))
+    {
+      in.setstate(std::ios::failbit);
+    }
+    return in;
+  }
+}
