+++ nikiforov.alexander/T2/delimiter.hpp
@@ -1,0 +1,15 @@
+#ifndef DELIMITER_HPP
+#define DELIMITER_HPP
+#include <sstream>
+
+namespace nikiforov
+{
+  struct DelimiterIO
+  {
+    char exp;
+  };
+
+  std::istream& operator>>(std::istream& in, DelimiterIO&& dest);
+}
+
+#endif
