+++ nikiforov.alexander/T2/typesKeys.cpp
@@ -1,23 +1,8 @@
 #include "typesKeys.hpp"
+#include "delimiter.hpp"
 
 namespace nikiforov
 {
-  std::istream& operator>>(std::istream& in, DelimiterIO&& dest)
-  {
-    std::istream::sentry sentry(in);
-    if (!sentry)
-    {
-      return in;
-    }
-    char c = '0';
-    in >> c;
-    if (in && (c != dest.exp))
-    {
-      in.setstate(std::ios::failbit);
-    }
-    return in;
-  }
-
   std::istream& operator>>(std::istream& in, UnsignedLongLongIO&& dest)
   {
     std::istream::sentry sentry(in);
